Separated logic code into separate library.
The library is undocumented for now, but exists in order to allow
benchmarking of the main logic without requiring a GUI window.
